added libgcc linker stub (no cross-compiler yet)

+++ arch/i686/include/arch/i8253.h
@@ -5,6 +5,6 @@
 #define PIT_TPS 14551
 
 void init_pit(void);
-uint32_t pit_get_ticks(void);
+uint64_t pit_get_ticks(void);
 
 #endif
